Progress on plugin refactor - Changed src/index.js to not use es6 imports.
