+++ package.json
@@ -5,7 +5,6 @@
   "main": "dist/index.js",
   "scripts": {
     "prepublish": "npm run test",
-    "pretest": "npm run build",
     "build": "rollup -c",
     "test": "nyc --reporter=html --reporter=text ava ./test/*.test.js -s",
     "coverage": "nyc report --reporter=text-lcov | coveralls"
